requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.2
fsspec==2025.12.0
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.5
langchain-google-genai==4.1.2
langsmith==0.5.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
orjson==3.11.5
packaging==25.0
pillow==12.0.0
postgrest==2.27.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.27.0
reportlab==4.4.7
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
websockets==15.0.1
Werkzeug==3.1.4
yarl==1.22.0
zstandard==0.25.0
